+++ examples/ProductionSketch/Shimmer.h
@@ -9,8 +9,13 @@ class Shimmer : public Pattern {
     //Shimmer variables
     uint8_t maxValue[LED_COUNT];
     uint8_t direction[LED_COUNT];
-    float value[LED_COUNT];
-    float death[LED_COUNT];
+    
+    // TODO: Loss of precision here affects output.
+//    float value[LED_COUNT];
+//    float death[LED_COUNT];
+    uint8_t value[LED_COUNT];
+    uint8_t death[LED_COUNT];
+    
     uint8_t stepSize;
     uint8_t ledMax;
